AdvertRepository revised for feign client requests

+++ order/src/main/java/com/abdulkadir/order/controller/OrderController.java
@@ -29,7 +29,7 @@ public class OrderController {
     }
 
     @GetMapping("/availableAdvertRights/{id}")
-    public int getAvailableAdvertRights(@PathVariable("id") Long userId) {
+    public long getAvailableAdvertRights(@PathVariable("id") Long userId) {
         return orderService.getAvailableAdvertRights(userId);
     }
